fix(ci): first-party line limit — extract provider recovery state views and split oversized test parts
