+++ tests/unit/service/test_provider_recovery_parts/test_provider_recovery_part_002.py
@@ -8,11 +8,8 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 from awf.adapters.provider_failures import (
-    AGENT_AUTH_FAILED,
     AGENT_IDLE_TIMEOUT,
     AGENT_PROVIDER_CAPACITY_EXHAUSTED,
-    AGENT_TIMEOUT,
-    classify_provider_failure,
 )
 from awf.api.schemas import WorkspaceCreateRequest
 from awf.db.enums import FailureReason, WorkspaceStatus
@@ -1282,314 +1279,3 @@ async def test_create_provider_recovery_attempt_row_retry_no_model(factory):
         assert result is not None
         assert result != "terminal"
         assert result.provider_recovery["action"] == "retry"
-
-
-class TestClassifyProviderFailureRoundTrip:
-    """Round-trip tests: realistic stderr shapes → correct
-    ProviderFailureClassification for each agent adapter."""
-
-    def test_codex_capacity_exhausted(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_PROVIDER_CAPACITY_EXHAUSTED,
-            stdout="",
-            stderr="MODEL_CAPACITY_EXHAUSTED Please try again later.",
-            provider="openai",
-            model="gpt-5.3-codex",
-        )
-        assert result is not None
-        assert result.failure_type == "capacity"
-        assert result.provider == "openai"
-        assert result.model == "gpt-5.3-codex"
-        assert result.retryable is True
-        assert result.fallback_allowed is True
-        assert result.reason_code == AGENT_PROVIDER_CAPACITY_EXHAUSTED
-
-    def test_codex_quota_exhausted(self) -> None:
-        result = classify_provider_failure(
-            reason_code=None,
-            stdout="",
-            stderr="RESOURCE_EXHAUSTED RetryableQuotaError Retry-After: 90",
-            provider="openai",
-            model="gpt-5.3-codex",
-        )
-        assert result is not None
-        assert result.failure_type == "quota"
-        assert result.provider == "openai"
-        assert result.retryable is True
-        assert result.retry_after_seconds == 90
-
-    def test_codex_auth_failed(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_AUTH_FAILED,
-            stdout="could not authenticate",
-            stderr="Manual authorization is required. Please run /login.",
-            provider="openai",
-            model="gpt-5.3-codex",
-        )
-        assert result is not None
-        assert result.failure_type == "auth"
-        assert result.provider == "openai"
-        assert result.retryable is True
-
-    def test_claude_capacity_exhausted(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_PROVIDER_CAPACITY_EXHAUSTED,
-            stdout="Anthropic server overloaded",
-            stderr="",
-            provider="anthropic",
-            model="sonnet",
-        )
-        assert result is not None
-        assert result.failure_type == "capacity"
-        assert result.provider == "anthropic"
-        assert result.model == "sonnet"
-        assert result.retryable is True
-
-    def test_claude_usage_limit(self) -> None:
-        result = classify_provider_failure(
-            reason_code=None,
-            stdout="",
-            stderr="You've hit your usage limit for claude. Switch to another model.",
-            provider="anthropic",
-            model="sonnet",
-        )
-        assert result is not None
-        assert result.failure_type == "usage_limit"
-        assert result.provider == "anthropic"
-
-    def test_claude_auth_failed(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_AUTH_FAILED,
-            stdout="error authenticating",
-            stderr="Could not authenticate. Please set an auth method.",
-            provider="anthropic",
-            model="sonnet",
-        )
-        assert result is not None
-        assert result.failure_type == "auth"
-        assert result.provider == "anthropic"
-
-    def test_gemini_capacity_exhausted(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_PROVIDER_CAPACITY_EXHAUSTED,
-            stdout="model_capacity_exhausted",
-            stderr="",
-            provider="google",
-            model="gemini-2.5-pro",
-        )
-        assert result is not None
-        assert result.failure_type == "capacity"
-        assert result.provider == "google"
-        assert result.model == "gemini-2.5-pro"
-        assert result.retryable is True
-
-    def test_gemini_quota_exhausted(self) -> None:
-        result = classify_provider_failure(
-            reason_code=None,
-            stdout="",
-            stderr=(
-                "429 Too Many Requests\n"
-                "google.api_core.exceptions.ResourceExhausted: 429 Quota exceeded"
-            ),
-            provider=None,
-            model="gemini-2.5-pro",
-        )
-        assert result is not None
-        assert result.failure_type == "quota"
-        assert result.provider == "google"
-
-    def test_gemini_auth_failed(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_AUTH_FAILED,
-            stdout="",
-            stderr="invalid_grant: Bad Request. Please check GEMINI_API_KEY.",
-            provider="google",
-            model="gemini-2.5-pro",
-        )
-        assert result is not None
-        assert result.failure_type == "auth"
-        assert result.provider == "google"
-
-    def test_gemini_usage_limit(self) -> None:
-        result = classify_provider_failure(
-            reason_code=None,
-            stdout="",
-            stderr="You've hit your usage limit. Switch to another model.",
-            provider="google",
-            model="gemini-2.5-pro",
-        )
-        assert result is not None
-        assert result.failure_type == "usage_limit"
-        assert result.provider == "google"
-
-    def test_opencode_ollama_capacity_exhausted(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_PROVIDER_CAPACITY_EXHAUSTED,
-            stdout="",
-            stderr="model_capacity_exhausted server overloaded",
-            provider="ollama",
-            model="deepseek-v4-pro",
-        )
-        assert result is not None
-        assert result.failure_type == "capacity"
-        assert result.provider == "ollama"
-        assert result.model == "deepseek-v4-pro"
-        assert result.retryable is True
-
-    def test_opencode_ollama_quota(self) -> None:
-        result = classify_provider_failure(
-            reason_code=None,
-            stdout="",
-            stderr="http 429 rate limit exceeded",
-            provider="ollama",
-            model="deepseek-v4-pro",
-        )
-        assert result is not None
-        assert result.failure_type == "quota"
-        assert result.provider == "ollama"
-
-    def test_opencode_ollama_auth_failed(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_AUTH_FAILED,
-            stdout="unauthorized",
-            stderr="ollama api key or ollama cloud authentication required.",
-            provider="ollama",
-            model="deepseek-v4-pro",
-        )
-        assert result is not None
-        assert result.failure_type == "auth"
-        assert result.provider == "ollama"
-
-    def test_timeout_with_provider_and_model(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_TIMEOUT,
-            stdout="",
-            stderr="Connection timed out after 600 seconds.",
-            provider="openai",
-            model="gpt-5.3-codex",
-        )
-        assert result is not None
-        assert result.failure_type == "timeout"
-        assert result.provider == "openai"
-        assert result.model == "gpt-5.3-codex"
-        assert result.retryable is True
-
-    def test_unclassified_output_returns_none(self) -> None:
-        result = classify_provider_failure(
-            reason_code=None,
-            stdout="SyntaxError",
-            stderr="unexpected token",
-            provider="openai",
-            model="gpt-5.3-codex",
-        )
-        assert result is None
-
-    def test_no_work_failure_is_classified_as_retryable(self) -> None:
-        result = classify_provider_failure(
-            reason_code=AGENT_PROVIDER_CAPACITY_EXHAUSTED,
-            stdout="",
-            stderr="MODEL_CAPACITY_EXHAUSTED no work was done",
-            provider="google",
-            model="gemini-2.5-pro",
-        )
-        assert result is not None
-        assert result.retryable is True
-        assert result.failure_type == "capacity"
-
-
-class TestFallbackInheritanceCompleteness:
-    """Prove every field in the fallback contract is inherited from source."""
-
-    @pytest.mark.unit
-    async def test_fallback_inherits_all_v2_fields_exhaustively(
-        self,
-        factory: async_sessionmaker[AsyncSession],
-    ) -> None:
-        service = WorkspaceService(factory)
-        source_response = await service.create(_request())
-
-        async with factory() as session:
-            repo = WorkspaceRepository(session)
-            source = await repo.get(source_response.id)
-            assert source is not None
-            await repo.transition(source, to=WorkspaceStatus.provisioning, reason_code="SEED")
-            source.branch_name = "awf/ws_old"
-            source.remote_push_branch = "awf/ws_old"
-            source.failure_reason = FailureReason.agent_failure.value
-            source.failure_message = "RESOURCE_EXHAUSTED RetryableQuotaError"
-            source.task_policy = {
-                **source.task_policy,
-                "provider_recovery_state": {"retry_attempt_number": 1},
-            }
-            await repo.transition(
-                source,
-                to=WorkspaceStatus.failed,
-                reason_code="AGENT_PROVIDER_CAPACITY_EXHAUSTED",
-                payload={
-                    "reason_code": "AGENT_PROVIDER_CAPACITY_EXHAUSTED",
-                    "message": source.failure_message,
-                    "details": {
-                        "provider": "google",
-                        "model": "gemini-2.5-pro",
-                        "retryable": True,
-                    },
-                },
-            )
-            await session.commit()
-
-        async with factory() as session:
-            result = await create_provider_recovery_attempt_row(
-                session,
-                source_response.id,
-                now=datetime(2026, 5, 1, 12, 0, tzinfo=UTC),
-            )
-            assert result is not None
-            await session.commit()
-            new_workspace_id = result.new_workspace_id
-
-        async with factory() as session:
-            repo = WorkspaceRepository(session)
-            source = await repo.get(source_response.id)
-            fallback = await repo.get(new_workspace_id)
-            attempts = list(
-                (
-                    await session.execute(
-                        select(TaskAttempt).order_by(TaskAttempt.attempt_number.asc())
-                    )
-                ).scalars()
-            )
-
-        assert source is not None
-        assert fallback is not None
-
-        assert fallback.repo_url == source.repo_url
-        assert fallback.branch_base == source.branch_base
-        assert fallback.task_title == source.task_title
-        assert fallback.task_prompt == source.task_prompt
-        assert fallback.task_external_id == source.task_external_id
-        assert fallback.task_class == source.task_class
-        assert fallback.owned_paths == source.owned_paths
-        assert fallback.test_commands == source.test_commands
-        assert fallback.profile_ref == source.profile_ref
-        assert fallback.requested_profile == source.requested_profile
-        assert fallback.resolved_profile == source.resolved_profile
-        assert fallback.auto_merge == source.auto_merge
-        assert fallback.initial_review_grace_period_seconds == (
-            source.initial_review_grace_period_seconds
-        )
-        assert fallback.task_kind == source.task_kind
-
-        assert fallback.task_policy is not None
-        assert isinstance(fallback.task_policy, dict)
-        recovery_state = fallback.task_policy.get("provider_recovery_state")
-        assert isinstance(recovery_state, dict)
-        assert recovery_state["source_workspace_id"] == source.id
-        assert recovery_state["fallback_attempt_number"] == 1
-        assert recovery_state["retry_attempt_number"] == 0
-        assert "source_canonical_attempt_id" not in recovery_state
-
-        assert len(attempts) >= 2
-        fallback_attempt = attempts[-1]
-        assert fallback_attempt.parent_attempt_id == attempts[0].id
-        assert fallback_attempt.redispatch_from_attempt_id == attempts[0].id
-        assert fallback_attempt.is_canonical_for_merge is False
